requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.5
django-crispy-forms==1.10.0
django-filter==2.4.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pycryptodome==3.9.9
PyJWT==2.0.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.4
razorpay==1.2.0
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
